Add idempotent and confirm flags
Idempotent option will only shift a given filesystem once. If a
filesystem appears to be owned by proper uid and gids, idmapshift
will exit early.

Confirm flag will only check to see if the filesystem is owned by
the proper uid and gids. Script will exit with code 1 if the
filesystem is not owned by the proper uid and gids.