Fix case where guest ID is greater than count
The code to determine if an ID was within a mapping range erroneously compared
the range (guest ID, count) instead of (guest ID, guest ID + count) which
caused mappings where guest ID was greater than count to not work.

For example, a mapping of 500:10000:50 would cause 501 to be mapped to NOBODY
instead of the correct 10001.

The fix is to use `guest ID + count` as the upper bound.

Fixes #3